Fixes some integer conversion macros

+++ Moira/Moira.h
@@ -213,7 +213,7 @@ protected:
     virtual void write16 (u32 addr, u16 val) = 0;
 
     // Provides the interrupt level in IRQ_USER mode
-    virtual int readIrqUserVector(u8 level) { return 0; }
+    virtual u16 readIrqUserVector(u8 level) { return 0; }
 
     // Instrution delegates
     virtual void signalReset() { };
@@ -350,7 +350,7 @@ private:
     void pollIrq() { reg.ipl = ipl; }
     
     // Selects the IRQ vector to branch to
-    int getIrqVector(int level);
+    u16 getIrqVector(int level);
     
 #include "MoiraInit.h"
 #include "MoiraALU.h"
